add basic support for snippets
A snippet is a pre-determined piece of code that can be added to one
or more runnables, specifically the run and setup steps. It can
take the executor settings and resources to render into the template,
and optionally can include the snakemake command. This first example
will use MPI with snakemake.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>