feat: show logs script
Problem: we need to be able to stream output logs to the console
Solution: Batch provide an API to do so, but the ratelimit is far too
low to be reasonable used in a multi-job context at a reasonable speed.
For the time being we are adding this as an example script, and need
to discuss a strategy to support this within the executor proper.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>