feat: add/googlebatch logging testing (#13)
This PR will add the following:

- an example script to show logs (not added to the executor per problems
with the API limit, described in
https://github.com/snakemake/snakemake-executor-plugin-googlebatch/issues/14
- ability for the true test in CI to take environment variables to
specify project /region (for easiest local running)
 - a missing NOTICE file required for HPCIC work

I'm working on an issue that describes the logging problem, and will be
sharing this with Google and will update here after. TBA.

Update: I don't see a solution that I like for logging. It either
requires enabling an extra service, or risking hitting an API limit. For
now we can easily see status updates in the UI and direct the interested
user to this script for dumping a log.

---------

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
Co-authored-by: vsoch <vsoch@users.noreply.github.com>