feat: add/intel mpi example (#20)
This adds the intel mpi example. The example functionally works, but the
actual command does not write the output to file (it is empty). I pinged
@johanneskoester about this and likely it needs to be debugged (the
workflow is also in core snakemake) but technically speaking the google
batch portion of this is working as expected so I think we can PR here.
My credits for this project are expiring early January (and I haven't
used a ton) so we need to keep moving.

---------

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
Co-authored-by: vsoch <vsoch@users.noreply.github.com>